feat: separate to web-server and db-service processes
