+++ config/nginx.conf
@@ -1,8 +1,8 @@
 events {}
 
 http {
-    upstream app {
-        server app:5000;
+    upstream web-server {
+        server web-server:5000;
     }
 
     server {
@@ -12,7 +12,7 @@ http {
             proxy_set_header Host $host;
             proxy_set_header X-Real-IP $remote_addr;
             proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
-            proxy_pass http://app;
+            proxy_pass http://web-server;
         }
     }
 }
