+++ database-service/requirements.txt
@@ -1,0 +1,9 @@
+# Third-party imports
+flask
+neo4j
+
+# OpenTelemetry imports
+opentelemetry-api
+opentelemetry-sdk
+opentelemetry-exporter-otlp-proto-grpc
+opentelemetry-instrumentation-flask
