+++ docker-compose.yml
@@ -7,10 +7,19 @@ services:
     volumes:
       - ./config/nginx.conf:/etc/nginx/nginx.conf:ro
     depends_on:
-      - app
+      - web-server
 
-  app:
-    build: ./python-example-app
+  web-server:
+    build: ./flask-web-server
+    deploy:
+      mode: replicated
+      replicas: 3
+    environment:
+      - "OTLP_ENDPOINT=tempo:4317"
+      - "DB_SERVICE_URL=http://db-service:5001"
+
+  db-service:
+    build: ./database-service
     deploy:
       mode: replicated
       replicas: 3
